Fix for Issue #1
Tempo2 fails to compute errors when using QR decomp
Bug is caused by assuming that double** cvm was a "blas" matrix, when in fact it is usually not so. Fixed by creating a temporary blas CVM and copying in the values as required. Should not impact performance.
Also fixed bug where chisq was not returned from TKfit