Improve consistency of START/FINISH flags
setting the fitFlag of START or FINISH to 2 now defines start/finish
based on bat. Setting it to 1 defines start/finish based on sat.

Previous behaviour was for the fit routines to use sats (and some other
bits of code) and e.g. plk use bats. This causes some strange unintended
behavior when zooming on plk.

textOutput and the like updated to interpret the 1/2 fitFlag correctly.

Soem plugins may be outdated, but mostly I think it will work fine


Former-commit-id: c79de5731217deee7eccd37ef0f85d0a9c7f5c60