Don't modify TZR parameters in transform. Fix for #402
The transform plugin used to convert the TZR parameters
as if they were barycentric MJD and freq. But they are
actually defined at TZRSITE, which is usually not the
barycentre. In any case, when converting a .par file the
TZR parameter seemed wrong until you re-fit the parfile
I am not totally certain, but I suspect that if TZRSITE is
the barycentre then it may indeed need correction, so I have
left it in for cases where the TZRSITE is @


Former-commit-id: 35c76f6ce77e751860dde4c2c3cc4257d5a8aa08